requirements.txt
# =========================
# Core API
# =========================
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
pyyaml>=6.0.2

# =========================
# AI / ML Stack (aligned)
# =========================
torch>=2.5.1,<2.6
transformers>=4.47.0
accelerate>=1.2.0
sentencepiece>=0.2.0
protobuf>=5.29.0

# =========================
# LLM Clients
# =========================
openai>=1.58.0
anthropic>=0.40.0

# =========================
# Database
# =========================
sqlalchemy[asyncio]>=2.0.36
psycopg2-binary>=2.9.10
asyncpg>=0.30.0
pymongo>=4.10.0
motor>=3.6.0
redis[hiredis]>=5.2.0
qdrant-client>=1.12.0
chromadb>=0.5.23
cachetools>=5.5.0
dependency-injector>=4.41.0

# =========================
# Security
# =========================
cryptography>=44.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# =========================
# Utilities
# =========================
requests>=2.32.3
beautifulsoup4>=4.12.3
numpy>=2.2.0
aiofiles>=24.1.0
httpx>=0.28.0
aiohttp>=3.11.0
tenacity>=9.0.0
psutil>=6.1.0
py-cpuinfo>=9.0.0
GPUtil>=1.4.0
rich>=13.9.0
click>=8.1.7
python-json-logger>=2.0.7
PyJWT>=2.10.0
slowapi>=0.1.9
strawberry-graphql[fastapi]>=0.219.2

# =========================
# Containers & Infra
# =========================
docker>=7.1.0
kubernetes>=31.0.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-sqlalchemy>=0.43b0
opentelemetry-instrumentation-redis>=0.43b0
opentelemetry-exporter-otlp>=1.22.0

# =========================
# Testing
# =========================
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
